requirements.txt
PyQt6>=6.5.0
mysql-connector-python>=8.1.0

